Client : ResourceManager 파일 분리
