+++ Client/Client/Client.vcxproj
@@ -149,6 +149,7 @@
     <ClInclude Include="Player.h" />
     <ClInclude Include="RenderConfig.h" />
     <ClInclude Include="Resource.h" />
+    <ClInclude Include="ResourceManager.h" />
     <ClInclude Include="RigidBody.h" />
     <ClInclude Include="Scene.h" />
     <ClInclude Include="Shader.h" />
@@ -180,6 +181,7 @@
     <ClCompile Include="OnlineScene.cpp" />
     <ClCompile Include="Player.cpp" />
     <ClCompile Include="RenderConfig.cpp" />
+    <ClCompile Include="ResourceManager.cpp" />
     <ClCompile Include="RigidBody.cpp" />
     <ClCompile Include="Scene.cpp" />
     <ClCompile Include="Shader.cpp" />
